feat: add folder migrations

+++ migrations/up/000002_create_table_products.up.sql
@@ -18,8 +18,6 @@ CREATE TABLE products (
 
 
 
-
-
 CREATE TABLE sizes (
     id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     name VARCHAR(10) NOT NULL UNIQUE,
@@ -39,6 +37,6 @@ CREATE TABLE products_sizes (
 
 CREATE TABLE products_images(
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    product_id INT REFERENCES products(id),
+    product_id BIGINT REFERENCES products(id),
     image VARCHAR(100)
 );
